cmd/catalog-repository/refresh: update time to int64

+++ cmd/cycloid/catalog-repositories/refresh.go
@@ -48,7 +48,7 @@ func refreshCatalogRepository(cmd *cobra.Command, args []string) error {
 		return err
 	}
 	fmt.Printf("id: %d    name: %s    url: %s    branch: %s    credential_id: %d\n", *cr.ID, *cr.Name, *cr.URL, cr.Branch, cr.CredentialID)
-	fmt.Printf("created_at: %v    updated_at: %v\n", time.Unix(*cr.CreatedAt, 0), time.Unix(*cr.UpdatedAt, 0))
+	fmt.Printf("created_at: %v    updated_at: %v\n", time.Unix(int64(*cr.CreatedAt), 0), time.Unix(int64(*cr.UpdatedAt), 0))
 
 	//TODO: Wait PR merged https://github.com/cycloidio/youdeploy-http-api/pull/2066
 	// output is not available yet
